feat: Implement permissions and add duplicate contact check
This commit introduces permission management to enhance security.
Additionally, it includes a duplicate contact check to prevent duplicate entries.

+++ backend/src/modules/contacts/contacts.service.ts
@@ -1,4 +1,6 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ConflictException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
+import { UserRole } from 'src/enums';
+import { ITokenUser } from 'src/interfaces';
 import { CreateContactDto } from './dto/create-contact.dto';
 import { UpdateContactDto } from './dto/update-contact.dto';
 import { Contact } from './entities/contact.entity';
@@ -8,27 +10,50 @@ import { ContactsRepository } from './repositories/contacts.repository';
 export class ContactsService {
   constructor(private contactsRepo: ContactsRepository) {}
 
+  private checkPermission(user: ITokenUser, user_id: string) {
+    if (user.id !== user_id && user.role !== UserRole.Admin) throw new ForbiddenException('Insufficient Permission');
+  }
+
+  private async checkDuplicated(userId: string, createContactDto: CreateContactDto) {
+    const duplicate = await this.contactsRepo.findDuplicate(userId, createContactDto);
+
+    if (duplicate) throw new ConflictException('Contact already exists');
+  }
+
   async create(userId: string, createContactDto: CreateContactDto) {
+    await this.checkDuplicated(userId, createContactDto);
+
     return await this.contactsRepo.create(userId, createContactDto);
   }
 
-  async findMany() {
-    return await this.contactsRepo.findMany();
+  async findMany(user: ITokenUser) {
+    if (user.role === UserRole.Admin) return await this.contactsRepo.findMany();
+
+    return await this.contactsRepo.findMany({ where: { user_id: user.id } });
   }
 
-  async findUnique(id: string) {
-    return await this.contactsRepo.findUnique(id);
-  }
-
-  async update(id: string, updateContactDto: UpdateContactDto) {
+  async findUnique(tokenUser: ITokenUser, id: string) {
     const contact: Contact = await this.contactsRepo.findUnique(id);
+
+    this.checkPermission(tokenUser, contact.user_id);
+    if (!contact) throw new NotFoundException(`Contact #${id} not found`);
+
+    return contact;
+  }
+
+  async update(tokenUser: ITokenUser, id: string, updateContactDto: UpdateContactDto) {
+    const contact: Contact = await this.contactsRepo.findUnique(id);
+
+    this.checkPermission(tokenUser, contact.user_id);
     if (!contact) throw new NotFoundException(`Contact #${id} not found`);
 
     return this.contactsRepo.update(id, updateContactDto);
   }
 
-  async remove(id: string) {
+  async remove(tokenUser: ITokenUser, id: string) {
     const contact: Contact = await this.contactsRepo.findUnique(id);
+
+    this.checkPermission(tokenUser, contact.user_id);
     if (!contact) throw new NotFoundException(`Contact #${id} not found`);
 
     return await this.contactsRepo.remove(id);
